fix babel macros
