+++ packages/app/src/sandbox/eval/transpilers/babel/worker/evaluate.js
@@ -8,7 +8,7 @@ import isESModule from 'sandbox/eval/utils/is-es-module';
 import evaluateCode from 'sandpack-core/lib/runner/eval';
 import detectOldBrowser from '@codesandbox/common/lib/detect-old-browser';
 
-import { resolve } from './utils/resolve';
+import { patchedResolve } from './utils/resolvePatch';
 import { getBabelTypes } from './utils/babelTypes';
 
 let cache = {};
@@ -66,7 +66,7 @@ export default function evaluate(
     }
 
     if (requirePath === 'resolve') {
-      return (...args) => resolve(...args);
+      return patchedResolve();
     }
 
     if (requirePath === 'babel-register') {
@@ -131,7 +131,7 @@ export default function evaluate(
 
     const resolvedPath =
       cachedPaths[dirName][requirePath] ||
-      resolve(requirePath, {
+      patchedResolve().sync(requirePath, {
         filename: path,
         extensions: ['.js', '.json'],
       });
@@ -203,7 +203,7 @@ export function evaluateFromPath(
   availablePlugins: Object,
   availablePresets: Object
 ) {
-  const resolvedPath = resolve(path, {
+  const resolvedPath = patchedResolve().sync(path, {
     filename: path,
     extensions: ['.js', '.json'],
   });
