+++ packages/app/src/sandbox/eval/transpilers/babel/worker/utils/macrosPatch.ts
@@ -1,12 +1,12 @@
 import * as macrosPlugin from 'babel-plugin-macros';
-import { resolve } from './resolve';
+import { patchedResolve } from './resolvePatch';
 
 // eslint-disable-next-line
 let finalExports = function m(babel, options) {
   return macrosPlugin(babel, {
     ...options,
     resolvePath(source, basePath) {
-      return resolve(source, {
+      return patchedResolve().sync(source, {
         filename: basePath + '/index',
       });
     },
