--- packages/app/src/sandbox/eval/transpilers/babel/worker/utils/resolve.ts
@@ -1,39 +1,0 @@
-import {
-  resolveSync,
-  IResolveOptionsInput,
-} from 'sandpack-core/lib/resolver/resolver';
-import gensync from 'gensync';
-
-/**
- * Patch 'resolve' to configure it to resolve esmodules by default. babel-plugin-macros goes
- * for the commonjs version, which we don't download anymore by default.
- */
-export function resolve(
-  specifier: string,
-  opts: Partial<IResolveOptionsInput>
-) {
-  const fs = global.BrowserFS.BFSRequire('fs');
-
-  const isFile = p => {
-    try {
-      const stats = fs.statSync(p);
-      return stats.isFile();
-    } catch (err) {
-      return false;
-    }
-  };
-
-  const resolvedPath = resolveSync(specifier, {
-    filename: '/index.js', // idk...
-    extensions: ['.js', '.mjs', '.json', '.ts', '.tsx'],
-    moduleDirectories: ['node_modules'],
-    ...opts,
-    isFile: gensync({
-      sync: p => isFile(p),
-    }),
-    readFile: gensync({
-      sync: p => fs.readFileSync(p, 'utf8'),
-    }),
-  });
-  return resolvedPath;
-}
