+++ packages/app/src/sandbox/eval/transpilers/babel/worker/utils/resolvePatch.ts
@@ -1,9 +1,30 @@
-import { resolve } from './resolve';
+import { resolver } from 'sandpack-core/lib/resolver/resolver';
+import gensync from 'gensync';
 
-function transformOptions(opts: any): any {
+function transformOptions(opts: any, fs: any): any {
+  const isFile = p => {
+    try {
+      const stats = fs.statSync(p);
+      return stats.isFile();
+    } catch (err) {
+      return false;
+    }
+  };
+
+  const basedir = opts.basePath || opts.basedir;
+  const filename = basedir ? basedir + '/index.js' : opts.filename;
   return {
-    ...opts,
-    filename: opts.basePath ? opts.basePath + '/index.js' : '/index.js',
+    filename: filename || '/index.js',
+    extensions: opts.extensions
+      ? opts.extensions
+      : ['.js', '.mjs', '.json', '.ts', '.tsx'],
+    moduleDirectories: ['node_modules'],
+    isFile: gensync({
+      sync: p => isFile(p),
+    }),
+    readFile: gensync({
+      sync: p => fs.readFileSync(p, 'utf8'),
+    }),
   };
 }
 
@@ -12,19 +33,17 @@ function transformOptions(opts: any): any {
  * for the commonjs version, which we don't download anymore by default.
  */
 export function patchedResolve() {
+  const fs = global.BrowserFS.BFSRequire('fs');
+
   const handler = {
     get(target, prop) {
-      if (prop === 'sync') {
-        return (p, options) => {
-          resolve(p, transformOptions(options));
-        };
-      }
-
-      // Useful for logging if somethings sync or not...
       return (p, options) => {
-        resolve(p, transformOptions(options));
+        const resolverOptions = transformOptions(options, fs);
+        const resolved = target.sync(p, resolverOptions);
+        return resolved;
       };
     },
   };
-  return new Proxy({}, handler);
+
+  return new Proxy(resolver, handler);
 }
