--- packages/app/src/sandbox/eval/utils/resolve-utils.ts
@@ -1,9 +1,0 @@
-import { PackageJSON } from '@codesandbox/common/lib/types';
-
-export const packageFilter = (p: PackageJSON) => {
-  if (p.module) {
-    p.main = p.module;
-  }
-
-  return p;
-};
